Data.Bin

Description

Tools for aggregating numeric values into a set of discrete bins according to some binning specification.

See withBinner for main usage information, and Bin for the main binned data type, and binFreq for a common usage example.

Specifying the binning

type BinView a b = forall p. Profunctor p => p b b -> p a a

A bidirectional "view" to transform the data type before binning.

See linView for a linear binning, and logView for a logarithmic binning.

This type is essentially Iso from the lens library, and any Iso' from lens can be used here. However, it is important that all of these represent monotonic isomorphisms.

linView :: BinView a a

Linear binning

logView :: Floating a => BinView a a

Logarithmic binning (smaller bins at lower levels, larger bins at higher levels).

data BinSpec (n :: Nat) a b

Specification of binning.

A BinSpec n a b will bin values of type a into n bins, according to a scaling in type b.

Constructor is meant to be used with type application syntax to indicate n, like BinSpec 5 0 10 linView@

Constructors

BS

Fields

bsMin :: a

lower bound of values

bsMax :: a

upper bound of values

bsView :: BinView a b

binning view

linBS :: forall n a. a -> a -> BinSpec n a a

Convenient constructor for a BinSpec for a linear scaling.

Meant to be used with type application syntax:

linBS @5 0 10

logBS :: forall n a. Floating a => a -> a -> BinSpec n a a

Convenient constructor for a BinSpec for a logarithmic scaling.

Meant to be used with type application syntax:

logBS @5 0 10

binSpecIntervals :: forall n a b. (KnownNat n, Fractional b) => BinSpec n a b -> Vector (n + 1) a

Generate a vector of the boundaries deriving the bins from a BinSpec. Can be useful for debugging.

Creating and manipulating bins

data Bin s n

A Bin s n is a single bin index out of n partitions of the original data set, according to a BinSpec represented by s.

All Bins with the same s follow the same BinSpec, so you can safely use binRange withBinner.

Actually has n + 2 partitions, since it also distinguishes values that are outside the BinSpec range.

type Binner s n a = a -> Bin s n

The type of a "binning function", given by withBinner. See withBinner for information on how to use.

withBinner :: (KnownNat n, RealFrac b) => BinSpec n a b -> (forall s. Reifies s (BinSpec n a b) => Binner s n a -> r) -> r

With a BinSpec, give a "binning function" that you can use to create bins within a continuation. The binning function is meant to be used with TypeApplications to specify how many bins to use:

withBinner myBinSpec $ toBin ->
    show (toBin 2.8523)

Inspecting bins

binFin :: Bin s n -> Maybe (Finite n)

Extract, potentially, the Bin index. Will return Nothing if the original value was outside the BinSpec range.

See binIx for a more specific version, which indicates if the original value was too high or too low.

binRange :: forall n a b s. (KnownNat n, Fractional b, Reifies s (BinSpec n a b)) => Bin s n -> (Maybe a, Maybe a)

Extract the minimum and maximum of the range indicabed by a given Bin.

A Nothing value indicates that we are outside of the normal range of the BinSpec, so is "unbounded" in that direction.

binMin :: forall n a b s. (KnownNat n, Fractional b, Reifies s (BinSpec n a b)) => Bin s n -> Maybe a

Extract the minimum of the range indicabed by a given Bin.

A Nothing value means that the original value was below the minimum limit of the BinSpec, so is "unbounded" in the lower direction.

binMax :: forall n a b s. (KnownNat n, Fractional b, Reifies s (BinSpec n a b)) => Bin s n -> Maybe a

Extract the maximum of the range indicabed by a given Bin.

A Nothing value means that the original value was above the maximum limit of the BinSpec, so is "unbounded" in the upper direction.

Showing bins

displayBin

Arguments

:: (KnownNat n, Fractional b, Reifies s (BinSpec n a b))

=> (a -> String)
how to display a value

-> Bin s n

-> String

Display the interval maintained by a Bin.

displayBinDouble

Arguments

:: (KnownNat n, Fractional b, Reifies s (BinSpec n Double b))

=> Int
number of decimal places to round

-> Bin s n

-> String

Display the interval maintained by a Bin, if the Bin contains a Double.

In-depth inspection

data Pointed a

Data type extending a value with an extra "minimum" and "maximum" value.

Constructors

Bot

PElem !a

Top

pElem :: Pointed a -> Maybe a

Extract the item from a Pointed if it is neither the extra minimum or maximum.

binIx :: Bin s n -> Pointed (Finite n)

A more specific version of binFin that indicates whether or not the value was too high or too low for the BinSpec range.

Untyped

data SomeBin a n

A SomeBin a n is Bin s n, except with the BinSpec s hidden. It's useful for returning out of withBinner.

It has useful Eq and Ord instances.

To be able to "unify" two Bins inside a SomeBin, use sameBinSpec to verify that the two SomeBins were created with the same BinSpec.

Constructors

(Fractional b, Reifies s (BinSpec n a b)) => SomeBin

Fields

getSomeBin :: Bin s n

sameBinSpec :: forall s t n a b p. (Reifies s (BinSpec n a b), Reifies t (BinSpec n a b), KnownNat n, Eq a, Fractional b) => p s -> p t -> Maybe (s :~: t)

Verify that the two reified BinSpec types refer to the same one, allowing you to use functions like == and compare on Bins that you get out of a SomeBin.

Handy use patterns

binFreq :: forall n t a s. Foldable t => Binner s n a -> t a -> Map (Bin s n) Int

Generate a histogram: given a container of as, generate a frequency map of how often values in a given discrete bin occurred.

xs :: [Double]
xs = [1..100]

main :: IO ()
main = withBinner (logBS @10 5 50) $ toBin ->
    mapM_ ((b, n) -> putStrLn (displayBin 4 b ++ "t" ++ show n))
  . M.toList
  $ binFreq toBin xs

(-inf .. 5.0000)        4
[5.0000 .. 6.2946)      2
[6.2946 .. 7.9245)      1
[7.9245 .. 9.9763)      2
[9.9763 .. 12.5594)     3
[12.5594 .. 15.8114)    3
[15.8114 .. 19.9054)    4
[19.9054 .. 25.0594)    6
[25.0594 .. 31.5479)    6
[31.5479 .. 39.7164)    8
[39.7164 .. 50.0000)    10
[50.0000 .. +inf)       51